+++ config/plasma.ts
@@ -57,7 +57,31 @@ export default <NetworkData>{
     },
     aprHandlers: {
         ybAprHandler: {
+            contract: {
+                calls: [
+                    {
+                        name: 'gearbox usdt0 pool',
+                        chain: 'PLASMA',
+                        contract: '0x76309a9a56309104518847bba321c261b7b4a43f',
+                        abi: 'function supplyRate() view returns(uint256)',
+                        functionName: 'supplyRate',
+                        parser: (rate) => Number(rate) * 10 ** -27,
+                        token: '0x76309a9a56309104518847bba321c261b7b4a43f',
+                    },
+                ],
+            },
             http: [
+                {
+                    url: 'https://app.avantprotocol.com/api/savusdApy',
+                    scale: 100,
+                    extractors: [
+                        {
+                            type: 'path',
+                            token: '0xa29420057f3e3b9512d4786df135da1674bd74d4',
+                            path: '$.savusdApy',
+                        },
+                    ],
+                },
                 {
                     url: 'https://api.maple.finance/v2/graphql',
                     body: JSON.stringify({